+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Allowed headings:
 
 ## [Unreleased]
 
+### Changed
+
+* `EditableListView` changed to have public `editMode` getter and `setIconButtonClickListener` method.
+
 ## [4.5.0] - 2023-02-17
 
 ### Changed